+++ libraries/AP_ServoRelayEvents/AP_ServoRelayEvents.cpp
@@ -112,7 +112,7 @@ bool AP_ServoRelayEvents::do_repeat_servo(uint8_t _channel, uint16_t _servo_valu
     case SRV_Channel::k_rcin1_mapped ... SRV_Channel::k_rcin16_mapped: {
         // mapped channels are set up with a -/+ 4500 angle range by
         // SRV_Channel::aux_servo_function_setup
-        int16_t angle_scaled = constrain_uint16(_servo_value, 1000, 2000);
+        int32_t angle_scaled = constrain_uint16(_servo_value, 1000, 2000);
         angle_scaled = (angle_scaled - 1500) * 9; // 1000 ... 2000 -> -500 ... 500 -> -4500 ... 4500
         _servo_value = c->pwm_from_scaled_value(angle_scaled);
         break;
